+++ Packages/GameCanvas/Runtime/Scripts/Engine/GcInputCameraEngine.cs
@@ -132,17 +132,14 @@ namespace GameCanvas.Engine
             var texture = GetOrCreateCameraTexture(camera, request);
             if (texture != null)
             {
-                resolution = new int2(texture.width, texture.height);
                 if (!texture.isPlaying)
                 {
                     texture.Play();
-                    return texture.isPlaying;
                 }
+                resolution = GetRotatedCameraSize(texture);
+                return texture.isPlaying;
             }
-            else
-            {
-                resolution = default;
-            }
+            resolution = default;
             return false;
         }
 
@@ -174,7 +171,7 @@ namespace GameCanvas.Engine
             texture = GetOrCreateCameraTexture(camera, request);
             if (texture != null)
             {
-                return new int2(texture.width, texture.height);
+                return GetRotatedCameraSize(texture);
             }
             return int2.zero;
         }
@@ -205,6 +202,7 @@ namespace GameCanvas.Engine
 
         public bool TryGetCameraImageAll(out ReadOnlyCollection<GcCameraDevice> array)
         {
+            InitCameraDevice();
             if (m_DeviceListReadOnly != null)
             {
                 array = m_DeviceListReadOnly;
@@ -229,7 +227,7 @@ namespace GameCanvas.Engine
         {
             if (m_TextureDict.TryGetValue(camera.DeviceName, out var texture))
             {
-                resolution = new int2(texture.width, texture.height);
+                resolution = GetRotatedCameraSize(texture);
                 return true;
             }
             resolution = default;
@@ -316,6 +314,50 @@ namespace GameCanvas.Engine
 
         void IEngine.OnBeforeUpdate(in System.DateTimeOffset now) { }
 
+        internal float2x3 CalcCameraMatrix(in WebCamTexture tex, in GcAnchor anchor)
+        {
+            var size = new float2(tex.width, tex.height);
+            var mtx = GcAffine.FromScale(size);
+            var offset = GcGraphicsEngine.GetOffset(anchor) - GcGraphicsEngine.GetOffset(GcAnchor.MiddleCenter);
+
+            if (tex.videoVerticallyMirrored)
+            {
+                var t = size * offset;
+                mtx = GcAffine.FromTranslate(t)
+                    .Mul(GcAffine.FromScale(new float2(1f, -1f)))
+                    .Mul(GcAffine.FromTranslate(-t))
+                    .Mul(mtx);
+            }
+
+            var deg = GcMath.Repeat(tex.videoRotationAngle, 360f);
+            if (GcMath.AlmostSame(deg, 90f) || GcMath.AlmostSame(deg, 270f))
+            {
+                mtx = GcAffine.FromTranslate(new float2(size.y, size.x) * offset)
+                    .Mul(GcAffine.FromRotate(math.radians(deg)))
+                    .Mul(GcAffine.FromTranslate(size * -offset))
+                    .Mul(mtx);
+            }
+            else if (GcMath.AlmostSame(deg, 180f))
+            {
+                var t = size * offset;
+                mtx = GcAffine.FromTranslate(t)
+                    .Mul(GcAffine.FromRotate(math.radians(deg)))
+                    .Mul(GcAffine.FromTranslate(-t))
+                    .Mul(mtx);
+            }
+
+            return mtx;
+        }
+
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        private static int2 GetRotatedCameraSize(in WebCamTexture texture)
+        {
+            var deg = Mathf.Repeat(-texture.videoRotationAngle, 360f);
+            return (GcMath.AlmostSame(deg, 90f) || GcMath.AlmostSame(deg, 270f))
+                ? new int2(texture.height, texture.width)
+                : new int2(texture.width, texture.height);
+        }
+
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         void InitCameraDevice()
         {
@@ -339,11 +381,12 @@ namespace GameCanvas.Engine
             }
             else
             {
-                if (callback != null)
-                {
-                    m_Context.Behaviour.OnFocusOnce += () => callback.Invoke(HasUserAuthorizedPermissionCamera);
-                }
+                var onFocus = false;
+                m_Context.Behaviour.OnFocusOnce += () => onFocus = true;
                 UnityEngine.Android.Permission.RequestUserPermission(UnityEngine.Android.Permission.Camera);
+                while (!onFocus) yield return null;
+                yield return null;
+                callback?.Invoke(HasUserAuthorizedPermissionCamera);
             }
 #elif UNITY_IOS
             if (HasUserAuthorizedPermissionCamera)