+++ Packages/GameCanvas/Sample~/DeviceCamera/Assets/GameCanvas/AutoGeneratedScripts/GcSound.cs
@@ -1,0 +1,17 @@
+/*------------------------------------------------------------*/
+// <summary>GameCanvas for Unity</summary>
+// <author>Seibe TAKAHASHI</author>
+// <remarks>
+// (c) 2015-2020 Smart Device Programming.
+// This software is released under the MIT License.
+// http://opensource.org/licenses/mit-license.php
+// </remarks>
+/*------------------------------------------------------------*/
+namespace GameCanvas
+{
+    public readonly partial struct GcSound : System.IEquatable<GcSound>
+    {
+        internal const int __Length__ = 0;
+
+    }
+}